[heap dump docs] Use nsenter to bypass systemd security protections
Since these instructions were first written, we have configured solr as a systemd service with `PrivateTmp=true`.  This is a good security protection to have, but it unfortunately also blocks the heap dump.

The nsenter command allows us to bypass this protection temporarily by entering the "namespace" of the solr process.

Also, format the steps as an ordered list.