[catalog] Include call_number_display field in the default fl
Orangelight has a feature where users can text themselves the call number of an item.
While working on pulibrary/orangelight#4479, we found that this feature wasn't working,
because we never actually told solr that we want to get the call_number_display field.

This adds call_number_display to the fl list, so it is retrieved by default.