[docker] Update path to run-initdb script in lando script
Previously, starting this container with lando with the lando-start.sh
command gave the error:

server/scripts/lando-start.sh: line 36: /opt/docker-solr/scripts/run-initdb: No such file or directory

The path to this script has changed slightly!

I confirmed that this script works with the updated path with:

1. In pul_solr: `docker build -t testing-solr-build docker`
1. In orangelight, `lando destroy -y`
1. In orangelight, update the image name for all solr containers to `testing-solr-build`
1. In orangelight, `bundle exec rake servers:start`

Thanks to @hackartisan for finding this issue!